Livingwind user direction (#228)
* Fix #173

Quickly wrapping up Save and Load.  This commit also fixes the bug when
pressing select all with no search results.

* Fix #215

Changed the default tab on load to be the planner and added a conditional
message directing the user to the Planning tab when no data is present.

* Fix #211

Selected/searched data now stays in the box after pressing the Plan
button.