Livingwind googlemaps (#229)
* Fix #173

Quickly wrapping up Save and Load.  This commit also fixes the bug when
pressing select all with no search results.

* Fix #215

Changed the default tab on load to be the planner and added a conditional
message directing the user to the Planning tab when no data is present.

* Fix #211

Selected/searched data now stays in the box after pressing the Plan
button.

* Map currently being displayed.  Sadly, I can't find a way to make
persistent changes work AND the map to load on tab swap at the same time.
This is most likely a rendering issue.

* Fix #224

Makes a google map showing the path of the trip.  The data is gathered
from the latitude and longitude property of the JSON object that gets
passed.

* Removed the SVG map from the site.

* Changed the styling of the map and added pins at each location.